Full screen everywhere.
